refactor: Reformat the code using prettier and fix linting issues
Prettier formatting was applied on the codebase along with linting fixes. Minor modifications were made to reflect the changes made for proper handling of server errors. A new error handler middleware `ErrorHandler` was added along with a package installation for class-validator. The MongoDB connect method for the database was changed to async/await to fix a deprecated code warning.